import hugo 2.11 preview
https://www.zeograd.com/download/hugo-2.11-preview.tar.gz

sha256sum hugo-2.11-preview.tar.gz
4b489911b6a44742cfb50d3578e01bda94127e26a84111b1adacdc6e05f8a0e0  hugo-2.11-preview.tar.gz

release notes:
Sources of the 2.11 preview version. Includes support for
netplay, arcade card, graphical configuration of keyboard/joypad, joypad
support under windows, zipped roms, hardware video rendering and some
more. Relies on gtk2, sdl, sdl-net developpement environment.

+++ cd.c
@@ -4,7 +4,7 @@
 unsigned
 Time2Frame (int min, int sec, int frame)
 {
-  return min * 60 * 75 + sec * 75 + frame;
+  return (unsigned)(min * 60 * 75 + sec * 75 + frame);
 }
 
 unsigned
@@ -16,25 +16,25 @@ Time2HSG (int min, int sec, int frame)
 unsigned
 Time2Redbook (int min, int sec, int frame)
 {
-  return ((min << 16) | (sec << 8) | (frame));
+  return (unsigned)((min << 16) | (sec << 8) | (frame));
 }
 
 void
 Frame2Time (unsigned frame, int *Min, int *Sec, int *Fra)
 {
-  *Min = (frame / (60 * 75));
+  *Min = (int)(frame / (60 * 75));
   frame -= *Min * 60 * 75;
-  *Sec = (frame / 75);
+  *Sec = (int)(frame / 75);
   frame -= *Sec * 75;
-  *Fra = frame;
+  *Fra = (int)frame;
 }
 
 void
 Redbook2Time (unsigned redbook, int *Min, int *Sec, int *Fra)
 {
-  *Fra = redbook & 0xff;
-  *Sec = (redbook >> 8) & 0xff;
-  *Min = (redbook >> 16) & 0xff;
+  *Fra = (int)(redbook & 0xff);
+  *Sec = (int)((redbook >> 8) & 0xff);
+  *Min = (int)((redbook >> 16) & 0xff);
 }
 
 void
